requirements.txt
requests==2.26.0
python-dotenv==0.21.1
setuptools==67.3.1
wheel==0.38.4
slick-bitcoinrpc==0.1.4
aiohttp==3.8.4
pylru==1.2.1
plyvel==1.5.0
websockets==10.4
python-rocksdb==0.7.0
